+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * add separator parameter to Environ.list
 * add cast parameter to Environ.list
+* add separator to Environ.tuple
 
 ## 0.3.1